Create contingency Table

contingencyTable creates a list of 2x2 contingency tables for Fisher's exact test.

contingencyTable(x, cols, signatures, features = "rownames")

Arguments

x

A SummarizedExperiment with rowData or a data frame.

cols

A character vector containing the names of the columns (must be logical) with the target (TRUE) and reference sets (FALSE).

signatures

A named list with signature sets.

features

A character vector of length 1. In case that x is a SummarizedExperiment, this argument will be used to create the features column. In case that x is a data frame, this argument must indicate the name of the column containing the features names (taxa, genes, etc.).

Value

A list of 2x2 contingenc tables (xtabs class).

See also

Examples

library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
contabs <- contingencyTable(x = se, cols = c("up", "down"), signatures = sigs)
contabs
#> $up_x_sig1
#>         contabXYZ.sig1
#> up       sig1 not_sig1
#>   up       13       17
#>   not_up  186      676
#> 
#> $up_x_sig2
#>         contabXYZ.sig2
#> up       sig2 not_sig2
#>   up       10       20
#>   not_up  315      547
#> 
#> $up_x_sig3
#>         contabXYZ.sig3
#> up       sig3 not_sig3
#>   up        7       23
#>   not_up  284      578
#> 
#> $down_x_sig1
#>           contabXYZ.sig1
#> down       sig1 not_sig1
#>   down        0       57
#>   not_down  199      636
#> 
#> $down_x_sig2
#>           contabXYZ.sig2
#> down       sig2 not_sig2
#>   down       38       19
#>   not_down  287      548
#> 
#> $down_x_sig3
#>           contabXYZ.sig3
#> down       sig3 not_sig3
#>   down        7       50
#>   not_down  284      551
#> 

df <- as.data.frame(SummarizedExperiment::rowData(se))
contabs2 <- contingencyTable(x = df, features = "taxa", cols = c("up", "down"), signatures = sigs)