Enrichment with Fisher's Exact Test (Hypergeometric)

enrichmentTestFisher performs a one-sided overrepresentation enrichment analysis using Fisher's exact test (hypergeometric). The function also calculates the odds.ratio using the epitools::oddsratio.wald function.

enrichmentTestFisher(contabs)

Arguments

contabs
A iist of contingency tables (xtabs objects); output from contingencyTable.

Value

A data frame; a report containing the p-value and odds ratio for each set x signature combination in the input list of contingency tables

See also

Examples

library(misctoolsr)
se <- exampleSE()
sigs <- exampleSignatures()
contabs <- contingencyTable(x = se, cols = c("up", "down"), signatures = sigs)
contabs

#> $up_x_sig1
#>         contabXYZ.sig1
#> up       sig1 not_sig1
#>   up       13       17
#>   not_up  186      676
#> 
#> $up_x_sig2
#>         contabXYZ.sig2
#> up       sig2 not_sig2
#>   up       10       20
#>   not_up  315      547
#> 
#> $up_x_sig3
#>         contabXYZ.sig3
#> up       sig3 not_sig3
#>   up        7       23
#>   not_up  284      578
#> 
#> $down_x_sig1
#>           contabXYZ.sig1
#> down       sig1 not_sig1
#>   down        0       57
#>   not_down  199      636
#> 
#> $down_x_sig2
#>           contabXYZ.sig2
#> down       sig2 not_sig2
#>   down       38       19
#>   not_down  287      548
#> 
#> $down_x_sig3
#>           contabXYZ.sig3
#> down       sig3 not_sig3
#>   down        7       50
#>   not_down  284      551
#> 

enrichment_result <- enrichmentTestFisher(contabs)

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.

enrichment_result

#> # A tibble: 6 x 7
#>   set   signature    p_value odds_ratio upper_ci lower_ci n_sig_genes
#>   <chr> <chr>          <dbl>      <dbl>    <dbl>    <dbl>       <dbl>
#> 1 down  sig2      0.00000162     3.77      6.58   2.16             38
#> 2 up    sig1      0.00732        2.80      5.73   1.37             13
#> 3 up    sig2      0.705          0.889     1.87   0.423            10
#> 4 up    sig3      0.907          0.649     1.46   0.288             7
#> 5 down  sig3      1.00           0.288     0.615  0.135             7
#> 6 down  sig1      1              0.0277    0.202  0.00382           0